changed middleware to provide headers and implemented auth for board

+++ src/app/api/board/route.js
@@ -2,9 +2,15 @@ import prisma from "@/lib/prisma";
 
 export async function GET(req) {
   try {
-    const cookie = req.cookies.get("user");
-    const userId = cookie.value;
+    const userId = req.headers.get("userId");
+    if (!userId) {
+      return Response.json({
+        success: false,
+        error: "Please log in to view boards",
+      });
+    }
     const boards = await prisma.board.findMany({
+      where: { userId },
       include: {
         columns: { include: { tasks: { include: { subtasks: true } } } },
       },
@@ -25,9 +31,17 @@ export async function POST(req) {
         error: "Please include a name for the board",
       });
     }
+    const userId = req.headers.get("userId");
+    if (!userId) {
+      return Response.json({
+        success: false,
+        error: "Please log in to view boards",
+      });
+    }
     const board = await prisma.board.create({
       data: {
         name,
+        userId,
       },
     });
     return Response.json({ success: true, board });
@@ -36,7 +50,8 @@ export async function POST(req) {
   }
 }
 
-export async function PUT(req) {
+// changes name of board
+export async function PATCH(req) {
   try {
     const data = await req.json();
     const { name, id } = data;
@@ -47,8 +62,20 @@ export async function PUT(req) {
           "Please include the ID of the existing board and what you want to name it",
       });
     }
-
     const board = await prisma.board.findUnique({ where: { id } });
+    const userId = req.headers.get("userId");
+    if (!userId) {
+      return Response.json({
+        success: false,
+        error: "Please log in to view boards",
+      });
+    }
+    if (userId !== board.userId) {
+      return Response.json({
+        success: false,
+        error: "You are not authorised to edit this board",
+      });
+    }
     if (board) {
       const newName = await prisma.board.update({
         where: { id },
@@ -77,6 +104,19 @@ export async function DELETE(req) {
       });
     }
     const isBoardExisting = await prisma.board.findUnique({ where: { id } });
+    const userId = req.headers.get("userId");
+    if (!userId) {
+      return Response.json({
+        success: false,
+        error: "Please log in to view boards",
+      });
+    }
+    if (isBoardExisting.userId !== userId) {
+      return Response.json({
+        success: false,
+        error: "Not authorised to delete this board",
+      });
+    }
     if (isBoardExisting) {
       const board = await prisma.board.delete({
         where: { id },
